requirements.txt
Flask==2.3.3
line-bot-sdk==2.4.1
gunicorn==21.2.0
